requirements.txt
numpy
torch>=1.1
numba
tensorboardX
easydict
pyyaml
scikit-image
tqdm
opencv-python
waymo-open-dataset-tf-2-1-0==1.2.0
